Move ID extraction into static methods
